[Aop done]

+++ src/main/java/com/bridgelabz/fundoonotes/model/Collaborater.java
@@ -17,16 +17,17 @@ public class Collaborater
 {
 @Id
 @GeneratedValue(strategy = GenerationType.AUTO)
-private int Id;
+private long Id;
 
 private String  email;
 @JsonIgnore
 @ManyToOne
 private Notes notes;
-public int getId() {
+
+public long getId() {
 	return Id;
 }
-public void setId(int id) {
+public void setId(long id) {
 	Id = id;
 }
 public String getEmail() {
